+++ lib/takeaway.rb
